feat: like

+++ src/utils/Api.js
@@ -56,18 +56,20 @@ class Api {
 
     }
 
-    putLike(cardId) {
-        return fetch(`${this._url}cards/likes/${cardId}`, {
-            method: "PUT",
-            headers: this._headers,
-        }).then(res => this._getServerData(res));
-    }
+    changeLikeCardStatus(cardId, isLiked) {
+        if (isLiked) {
+            return fetch(`${this._url}cards/likes/${cardId}`, {
+                method: "PUT",
+                headers: this._headers,
+            }) .then(res => this._getServerData(res));
+        }
+        else {
+            return fetch(`${this._url}cards/likes/${cardId}`, {
+                method: "DELETE",
+                headers: this._headers
+            }) .then(res => this._getServerData(res));
+        }
 
-    deleteLike(cardId) {
-        return fetch(`${this._url}cards/likes/${cardId}`, {
-            method: "DELETE",
-            headers: this._headers
-        }).then(res => this._getServerData(res));
     }
 
     deleteCard(cardId) {
